VSP recommended bug fixes: - Updated example scenes. - Documentation update - Added #if UNITY_EDITOR preprocessor to a UI function - Added project curves to scene view direction to make projecting curves more UX friendly. - Additionally, restarted pipe sessions will now create a pipe session instead of a socket one - Added some error checking with shelves - Additionally, fixed an issue where the Add selection menu wouldn't update when cooking is set to false. - Miscellaneous bug fixes regarding errors and UI. - Fixed minor undo error log

+++ Plugins/HoudiniEngineUnity/Scripts/Asset/HEU_AssetUpdater.cs
@@ -186,11 +186,12 @@ namespace HoudiniEngineUnity
 				"Prefab: " + prefabPath;
 
 		heu_root._houdiniAsset.WarnedPrefabNotSupported = true;
-		if (HEU_EditorUtility.DisplayDialog(title, message, "Remove Prefab & Revert", "Keep Prefab"))
+		if (HEU_EditorUtility.DisplayDialog(title, message, "Remove Prefab && Revert", "Keep Prefab"))
 		{
 		    HEU_EditorUtility.DisconnectPrefabInstance(instance);
-
 		    HEU_AssetDatabase.DeleteAssetAtPath(prefabPath);
+
+		    HEU_AssetDatabase.SaveAndRefreshDatabase();
 		}
 	    }
 #endif
